finish most works and UI updates

+++ app/src/main/java/top/kncweb/sposocapp/remote/api/UserInfoApiService.java
@@ -14,14 +14,14 @@ import top.kncweb.sposocapp.remote.models.UserInfo;
 public interface UserInfoApiService {
 
     @GET("userinfo/{id}")
-    Call<UniResponse<UserInfo>> getUserInfo(@Path("id") int uid);
+    Call<UniResponse<UserInfo>> getUserInfo(@Path("id") long uid);
 
     @POST("userinfo")
     Call<UniResponse<Integer>> addUserInfo(@Body UserInfo userInfo);
 
     @FormUrlEncoded
     @PUT("userinfo/{id}")
-    Call<UniResponse<Boolean>> updateUserInfo(@Path("id") int id,
+    Call<UniResponse<Boolean>> updateUserInfo(@Path("id") long id,
                                               @Field("height") Integer height,  // 使用 Integer 允许 null
                                               @Field("weight") Integer weight,
                                               @Field("sex") String sex);
